+++ src/glicko.rs
@@ -7,6 +7,7 @@ use serde::{Deserialize, Serialize};
 
 /// ln 10 / 400
 const Q: f64 = 0.005_756_5;
+pub const CONFIDENCE_INTERVAL_95: f64 = 1.96;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct Rating {
@@ -19,6 +20,15 @@ impl Rating {
     #[must_use]
     pub fn rd_sq(&self) -> f64 {
         self.rd * self.rd
+    }
+
+    #[must_use]
+    pub fn to_string_rounded(&self) -> String {
+        format!(
+            "{} ± {}",
+            self.rating.round(),
+            (CONFIDENCE_INTERVAL_95 * self.rd).round()
+        )
     }
 
     /// This assumes 30 2 month periods must pass before one's rating
@@ -64,9 +74,10 @@ impl Default for Rating {
 }
 
 impl fmt::Display for Rating {
-    // With a confidence interval of 95%.
+    // Note: We use a FIGURE SPACE before and after the ± so
+    // .split_ascii_whitespace() does not treat it as a space.
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}±{:.2}", self.rating.round(), 1.96 * self.rd.round())
+        write!(f, "{} ± {}", self.rating, CONFIDENCE_INTERVAL_95 * self.rd)
     }
 }